fjerner hendelseDbo
