+++ src/main/kotlin/no/nav/amt/distribusjon/hendelse/HendelseConsumer.kt
@@ -3,8 +3,6 @@ package no.nav.amt.distribusjon.hendelse
 import com.fasterxml.jackson.module.kotlin.readValue
 import no.nav.amt.distribusjon.Environment
 import no.nav.amt.distribusjon.application.plugins.objectMapper
-import no.nav.amt.distribusjon.hendelse.db.HendelseDbo
-import no.nav.amt.distribusjon.hendelse.db.HendelseRepository
 import no.nav.amt.distribusjon.hendelse.model.Hendelse
 import no.nav.amt.distribusjon.journalforing.JournalforingService
 import no.nav.amt.distribusjon.kafka.Consumer
@@ -16,7 +14,6 @@ import no.nav.amt.distribusjon.varsel.VarselService
 import org.apache.kafka.common.serialization.StringDeserializer
 import org.apache.kafka.common.serialization.UUIDDeserializer
 import org.slf4j.LoggerFactory
-import java.time.LocalDateTime
 import java.util.UUID
 
 class HendelseConsumer(
@@ -41,16 +38,7 @@ class HendelseConsumer(
     override suspend fun consume(key: UUID, value: String) {
         val hendelse: Hendelse = objectMapper.readValue(value)
         log.info("Mottatt hendelse ${hendelse.id} for deltaker ${hendelse.deltaker.id}")
-        hendelseRepository.insert(
-            HendelseDbo(
-                id = hendelse.id,
-                deltakerId = hendelse.deltaker.id,
-                deltaker = hendelse.deltaker,
-                ansvarlig = hendelse.ansvarlig,
-                payload = hendelse.payload,
-                opprettet = LocalDateTime.now(),
-            ),
-        )
+        hendelseRepository.insert(hendelse)
         varselService.handleHendelse(hendelse)
         journalforingService.handleHendelse(hendelse)
     }
